Rename bomb exploded event

+++ Bomberman/BaseCollider.cpp
@@ -78,7 +78,7 @@ void bomberman::BaseCollider::Render() const
 	//glm::vec3 color = { 1.0f, 0.0f, 0.0f }; // Red color
 	auto& renderer = dae::Renderer::GetInstance();
 
-	for (int vert = 0; vert < m_Polygon.size(); vert++)
+	for (int vert = 0; vert < static_cast<int>(m_Polygon.size()); vert++)
 	{
 		glm::vec2 p1 = m_Polygon[vert];
 		glm::vec2 p2 = m_Polygon[(vert + 1) % m_Polygon.size()];
